requirements.txt
langchain_core==1.1.3
langchain_openai==1.1.2
langgraph==1.0.4
Markdown==3.10
python-dotenv==1.2.1
Requests==2.32.5
config==0.5.1